+++ sdk/appservice/azure-resourcemanager-appservice/src/main/java/com/azure/resourcemanager/appservice/package-info.java
@@ -4,6 +4,6 @@
 
 /**
  * Package containing the classes for WebSiteManagementClient.
- * WebSite Management Client.
+ * AppService Management Client.
  */
 package com.azure.resourcemanager.appservice;