+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/implementation/Exceptions.java
@@ -70,4 +70,8 @@ public class Exceptions {
         return Exceptions.isStatusCode(cosmosException, HttpConstants.StatusCodes.GONE)
             && Exceptions.isSubStatusCode(cosmosException, HttpConstants.SubStatusCodes.LEASE_NOT_FOUND);
     }
+
+    public static boolean isClientAssignedSubStatusCodeForInternalServerError(int statusCode, int subStatusCode) {
+        return statusCode == HttpConstants.StatusCodes.INTERNAL_SERVER_ERROR && (subStatusCode >= 20_000 && subStatusCode < 21_000);
+    }
 }