+++ sdk/search/azure-search-documents/src/main/java/com/azure/search/documents/util/SearchPagedIterable.java
@@ -13,7 +13,6 @@ import com.azure.core.util.paging.ContinuablePagedIterable;
 import com.azure.core.util.paging.PageRetrieverSync;
 import com.azure.search.documents.implementation.models.SearchFirstPageResponseWrapper;
 import com.azure.search.documents.implementation.models.SearchRequest;
-import com.azure.search.documents.implementation.util.SemanticSearchResultsAccessHelper;
 import com.azure.search.documents.models.DebugInfo;
 import com.azure.search.documents.models.FacetResult;
 import com.azure.search.documents.models.SearchResult;
@@ -31,7 +30,10 @@ public final class SearchPagedIterable extends PagedIterableBase<SearchResult, S
      * Creates an instance of {@link SearchPagedIterable}.
      *
      * @param pagedFlux The {@link SearchPagedFlux} that will be consumed as an iterable.
+     * @deprecated Use {@link SearchPagedIterable#SearchPagedIterable(Supplier)} or
+     * {@link SearchPagedIterable#SearchPagedIterable(Supplier, Function)}.
      */
+    @Deprecated
     public SearchPagedIterable(SearchPagedFlux pagedFlux) {
         super(pagedFlux);
         this.pagedFlux = pagedFlux;
@@ -86,7 +88,10 @@ public final class SearchPagedIterable extends PagedIterableBase<SearchResult, S
      *
      * @return The percentage of the index covered in the search request if {@code minimumCoverage} was set in the
      * request, otherwise {@code null}.
+     * @deprecated Use {@link SearchPagedResponse#getCoverage()} when consuming {@link #streamByPage()} or
+     * {@link #iterableByPage()}.
      */
+    @Deprecated
     public Double getCoverage() {
         return metadataSupplier != null
             ? metadataSupplier.get().getFirstPageResponse().getCoverage()
@@ -99,7 +104,10 @@ public final class SearchPagedIterable extends PagedIterableBase<SearchResult, S
      * If {@code facets} weren't supplied in the request this will be {@code null}.
      *
      * @return The facet query results if {@code facets} were supplied in the request, otherwise {@code null}.
+     * @deprecated Use {@link SearchPagedResponse#getFacets()} when consuming {@link #streamByPage()} or
+     * {@link #iterableByPage()}.
      */
+    @Deprecated
     public Map<String, List<FacetResult>> getFacets() {
         return metadataSupplier != null
             ? metadataSupplier.get().getFirstPageResponse().getFacets()
@@ -113,7 +121,10 @@ public final class SearchPagedIterable extends PagedIterableBase<SearchResult, S
      *
      * @return The approximate number of documents that match the request if {@code count} is {@code true}, otherwise
      * {@code null}.
+     * @deprecated Use {@link SearchPagedResponse#getCount()} when consuming {@link #streamByPage()} or
+     * {@link #iterableByPage()}.
      */
+    @Deprecated
     public Long getTotalCount() {
         return metadataSupplier != null
             ? metadataSupplier.get().getFirstPageResponse().getCount()
@@ -127,10 +138,13 @@ public final class SearchPagedIterable extends PagedIterableBase<SearchResult, S
      *
      * @return The semantic search results if semantic search was requested, otherwise an empty
      * {@link SemanticSearchResults}.
+     * @deprecated Use {@link SearchPagedResponse#getSemanticResults()} when consuming {@link #streamByPage()} or
+     * {@link #iterableByPage()}.
      */
+    @Deprecated
     public SemanticSearchResults getSemanticResults() {
         return metadataSupplier != null
-            ? SemanticSearchResultsAccessHelper.create(metadataSupplier.get().getFirstPageResponse())
+            ? metadataSupplier.get().getFirstPageResponse().getSemanticResults()
             : pagedFlux.getSemanticResults().block();
     }
 
@@ -138,7 +152,10 @@ public final class SearchPagedIterable extends PagedIterableBase<SearchResult, S
      * The debug information that can be used to further explore your search results.
      *
      * @return The debug information that can be used to further explore your search results.
+     * @deprecated Use {@link SearchPagedResponse#getDebugInfo()} when consuming {@link #streamByPage()} or
+     * {@link #iterableByPage()}.
      */
+    @Deprecated
     public DebugInfo getDebugInfo() {
         return metadataSupplier != null
             ? metadataSupplier.get().getFirstPageResponse().getDebugInfo()